checkpoint document module
