+++ bokeh/document/__init__.py
@@ -1,0 +1,10 @@
+'''
+
+
+'''
+
+from .document import Document ; Document
+
+from .document import DEFAULT_TITLE ; DEFAULT_TITLE
+
+from .locking import without_document_lock ; without_document_lock
