+++ bokeh/document/tests/test_document.py
@@ -1,16 +1,16 @@
 from __future__ import absolute_import, print_function
 
-import pytest
 import unittest
 
 from copy import copy
 
-import bokeh.document as document
+import bokeh.document.document as document
+
 from bokeh.io import curdoc
 from bokeh.model import Model
 from bokeh.models import ColumnDataSource
 from bokeh.core.properties import Int, Instance, String, DistanceSpec
-from bokeh.protocol.events import (ColumnsPatchedEvent, ColumnsStreamedEvent, ModelChangedEvent, RootAddedEvent,
+from bokeh.document.events import (ColumnsPatchedEvent, ColumnsStreamedEvent, ModelChangedEvent, RootAddedEvent,
                                    RootRemovedEvent, SessionCallbackAdded, SessionCallbackRemoved, TitleChangedEvent)
 from bokeh.protocol.messages.patch_doc import process_document_events
 
@@ -966,47 +966,3 @@ class TestDocument(unittest.TestCase):
         self.assertEqual(2, len(d2.roots))
         self.assertEqual(42, d2.roots[0].foo)
         self.assertEqual(57, d2.roots[1].foo)
-
-
-class Test_UnlockedDocumentProxy(unittest.TestCase):
-
-    def test_next_tick_callback_works(self):
-        d = document._UnlockedDocumentProxy(document.Document())
-        assert curdoc() is not d
-        curdoc_from_cb = []
-        def cb():
-            curdoc_from_cb.append(curdoc())
-        callback = d.add_next_tick_callback(cb)
-        callback.callback()
-        assert len(curdoc_from_cb) == 1
-        assert curdoc_from_cb[0] is d._doc
-        def cb2(): pass
-        callback = d.add_next_tick_callback(cb2)
-        d.remove_next_tick_callback(cb2)
-
-    def test_other_attrs_raise(self):
-        d = document._UnlockedDocumentProxy(document.Document())
-        assert curdoc() is not d
-        with pytest.raises(RuntimeError) as e:
-            d.foo
-            assert str(e) == "Only add_next_tick_callback may be used safely without taking the document lock; "
-            "to make other changes to the document, add a next tick callback and make your changes "
-            "from that callback."
-        for attr in dir(d._doc):
-            if attr in ["add_next_tick_callback", "remove_next_tick_callback"]: continue
-            with pytest.raises(RuntimeError) as e:
-                getattr(d, "foo")
-
-    def test_without_document_lock(self):
-        d = document.Document()
-        assert curdoc() is not d
-        curdoc_from_cb = []
-        @document.without_document_lock
-        def cb():
-            curdoc_from_cb.append(curdoc())
-        callback = d.add_next_tick_callback(cb)
-        callback._callback()
-        assert callback.callback.nolock == True
-        assert len(curdoc_from_cb) == 1
-        assert curdoc_from_cb[0]._doc is d
-        assert isinstance(curdoc_from_cb[0], document._UnlockedDocumentProxy)
